Group 1
The son (player), who was originally addicted to his cell phone, neglected his mother's concern when he came home for a dinner party after a long time, but unexpectedly found himself in a strange world when he looked up. He saw the extreme events that happened in the streets of the other world due to cell phone addiction.